chore: Move used code from Win32Hook to MSWindowsHook and remove Win32Hook
